fix: 页面刷新请求问题
